requirements.txt
asgiref==3.3.0
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
Django==3.1.3
django-cors-headers==3.5.0
django-model-utils==4.0.0
djangorestframework==3.12.2
firebase-admin==4.4.0
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
httplib2==0.18.1
idna==2.10
msgpack==1.0.0
protobuf==3.13.0
psycopg2==2.8.6
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pytz==2020.4
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
Unipath==1.1
uritemplate==3.0.1
urllib3==1.25.11
